feature: fixing types and updating readme :sparkles:

+++ app/profile/page.tsx
@@ -4,10 +4,11 @@ import { useSession } from 'next-auth/react';
 import { useEffect, useState } from 'react';
 import { useRouter } from 'next/navigation';
 import { Profile } from '@components';
+import { Post } from '@types';
 
 const MyProfile = () => {
   const router = useRouter();
-  const { data: session } = useSession();
+  const { data: session } = useSession() as any;
 
   const [myPosts, setMyPosts] = useState([]);
 
@@ -22,20 +23,20 @@ const MyProfile = () => {
     if (session?.user.id) fetchPosts();
   }, [session?.user.id]);
 
-  const handleEdit = (post) => {
+  const handleEdit = (post: Post) => {
     router.push(`/update-prompt?id=${post._id}`);
   };
 
-  const handleDelete = async (post) => {
+  const handleDelete = async (post: Post) => {
     const hasConfirmed = confirm('Are you sure you want to delete this prompt?');
 
-    if (hasConfirmed) {
+    if (hasConfirmed && post?._id) {
       try {
-        await fetch(`/api/prompt/${post._id.toString()}`, {
+        await fetch(`/api/prompt/${post?._id.toString()}`, {
           method: 'DELETE'
         });
 
-        const filteredPosts = myPosts.filter((item) => item._id !== post._id);
+        const filteredPosts = myPosts.filter((item: Post) => item?._id !== post._id);
 
         setMyPosts(filteredPosts);
       } catch (error) {
